+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.2
+* Export internal widgets to simplify package reference
+* Update show picker add title
+* Update TextCell add hasUneventRows property
+
 ## 1.0.1
 * Add example and update descriptions/comments